test(api): split 4 oversized ws/trpc test files (api-4cmq)
Split four test files in apps/api/ (ws + tRPC layer) so every resulting
file is under 500 LOC, all tests preserved, no behavior changes.

- __tests__/ws/handlers.test.ts (1042) -> doc-ops + subscribe + submit
- ws/__tests__/handlers.test.ts (928) -> submit-errors + misc-errors
- __tests__/ws/message-router.test.ts (1406) -> auth + access + errors
- __tests__/trpc/routers/structure.test.ts (875) -> entity-type +
  entity + relations

The two handlers.test.ts files are not duplicates: the one under
__tests__/ws/ is the happy-path suite (real EncryptedRelay), the one
under ws/__tests__/ is the branch-coverage suite (mocked relay,
verification edge cases). Each was split independently with concern-
named files in its original directory.

Test counts preserved per source: 35+35+51+45 = 166 tests retained
across the splits.

Shared fixtures extracted (no vi.mock blocks; vi.mock stays per file
to keep hoisting correct):

- __tests__/helpers/ws-handlers-fixtures.ts
- ws/__tests__/handlers-fixtures.ts
- __tests__/ws/message-router-fixtures.ts
- __tests__/trpc/routers/structure-fixtures.ts

DRY consolidations:

- addAuthedConnection collapses ~20 lines of register+authenticate
  boilerplate per use site.
- makeBrokenRelayContext collapses inline broken-relay literals used
  by 6 INTERNAL_ERROR tests.
- makeSnapshotPayload consolidates the snapshot-envelope literal used
  by 4 SubmitSnapshotRequest tests.
